+++ pom.xml
@@ -51,6 +51,10 @@
       <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
       <version>2.3.0</version>
     </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-validation</artifactId>
+    </dependency>
 
     <!--3rd party library-->
     <dependency>
@@ -72,6 +76,7 @@
       <version>4.11.0</version>
       <scope>test</scope>
     </dependency>
+
 
   </dependencies>
