+++ src/main/java/it/gov/pagopa/register/dto/role/PortalConsentDTO.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.register.dto.rolepermission;
+package it.gov.pagopa.register.dto.role;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
