+++ src/main/java/it/gov/pagopa/register/exception/role/VersionNotMatchedException.java
@@ -1,13 +1,13 @@
-package it.gov.pagopa.register.exception.rolepermission;
+package it.gov.pagopa.register.exception.role;
 
 import it.gov.pagopa.common.web.exception.ServiceException;
 import it.gov.pagopa.common.web.exception.ServiceExceptionPayload;
-import it.gov.pagopa.register.constants.RolePermissionConstants;
+import it.gov.pagopa.register.constants.RoleConstants;
 
 public class VersionNotMatchedException extends ServiceException {
 
     public VersionNotMatchedException(String message) {
-        this(RolePermissionConstants.ExceptionCode.VERSION_NOT_MATCHED, message);
+        this(RoleConstants.ExceptionCode.VERSION_NOT_MATCHED, message);
     }
 
     public VersionNotMatchedException(String code, String message) {
