+++ src/main/java/it/gov/pagopa/register/repository/role/PortalConsentRepository.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.register.repository.rolepermission;
+package it.gov.pagopa.register.repository.role;
 
-import it.gov.pagopa.register.model.rolepermission.PortalConsent;
+import it.gov.pagopa.register.model.role.PortalConsent;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface PortalConsentRepository extends MongoRepository<PortalConsent, String> {
