+++ src/main/java/it/gov/pagopa/register/repository/role/RolePermissionRepository.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.register.repository.rolepermission;
+package it.gov.pagopa.register.repository.role;
 
-import it.gov.pagopa.register.model.rolepermission.RolePermission;
+import it.gov.pagopa.register.model.role.RolePermission;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 import java.util.Optional;
