+++ src/main/java/it/gov/pagopa/register/service/role/PortalConsentServiceImpl.java
@@ -1,12 +1,12 @@
-package it.gov.pagopa.register.service.rolepermission;
+package it.gov.pagopa.register.service.role;
 
 import it.gov.pagopa.register.connector.onetrust.OneTrustRestService;
-import it.gov.pagopa.register.dto.rolepermission.PortalConsentDTO;
-import it.gov.pagopa.register.dto.mapper.rolepermission.PrivacyNotices2PortalConsentDTOMapper;
+import it.gov.pagopa.register.dto.role.PortalConsentDTO;
+import it.gov.pagopa.register.dto.mapper.role.PrivacyNotices2PortalConsentDTOMapper;
 import it.gov.pagopa.register.dto.onetrust.PrivacyNoticesDTO;
-import it.gov.pagopa.register.exception.rolepermission.VersionNotMatchedException;
-import it.gov.pagopa.register.model.rolepermission.PortalConsent;
-import it.gov.pagopa.register.repository.rolepermission.PortalConsentRepository;
+import it.gov.pagopa.register.exception.role.VersionNotMatchedException;
+import it.gov.pagopa.register.model.role.PortalConsent;
+import it.gov.pagopa.register.repository.role.PortalConsentRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
